Moves implementation into a separate cpp file, allowing inclusion in multiple translation units
