+++ doxygen.config
@@ -1,6 +1,6 @@
 PROJECT_NAME = Configuru
 OUTPUT_DIRECTORY = docs
-INPUT = README.md configuru.hpp
+INPUT = README.md configuru.hpp configuru.cpp
 USE_MDFILE_AS_MAINPAGE = README.md
 RECURSIVE = NO
 WARN_IF_UNDOCUMENTED = NO
