+++ test/configuru_test.cpp
@@ -16,6 +16,7 @@
 
 #define CONFIGURU_IMPLEMENTATION 1
 #include <../configuru.hpp>
+#include <../configuru.cpp>
 
 // ----------------------------------------------------------------------------
